requirements-cpu.txt
opencv-python
numpy
scikit-learn
pyproj
pillow
filterpy
imutils
flask
